add all item icons, fix other item related bugs

+++ Assets/Scripts/inventory/OriItem.cs
@@ -1,8 +1,8 @@
 ﻿public class OriItem : Item
 {
-    public int multiplier { get; private set; }
+    public float multiplier { get; private set; }
     
-    public OriItem(string name, string description, string iconPath, int multiplier) : base(ItemType.Ori, name, description, iconPath)
+    public OriItem(string name, string description, string iconPath, float multiplier) : base(ItemType.Ori, name, description, iconPath)
     {
         this.multiplier = multiplier;
     }
